pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = ""
authors = ["Ayush Agarwal <dearayush2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
discord = "^2.3.2"
openai = "^1.23.6"
python-dotenv = "^1.0.1"
google-generativeai = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
